Implementada injeção de dependencias, JPA not work

+++ demo1/src/main/java/com/tf/demo1/Domínio/ItemPedido.java
@@ -1,18 +1,28 @@
 package com.tf.demo1.Domínio;
 
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+
+//@Entity
 public class ItemPedido {
 
-	private int codProduto;
+	//@Id
+	//@GeneratedValue(strategy=GenerationType.AUTO)
+	private Long codProduto;
 	private int quantidade;
 	private double preco;
 
-	public ItemPedido(int codProduto, int quantidade, double preco){
+	public ItemPedido(Long codProduto, int quantidade, double preco){
 		this.codProduto = codProduto;
 		this.quantidade = quantidade;
 		this.preco = preco;
 	}
 
-	public int getCodProduto() {
+	//protected ItemPedido() {}
+
+	public Long getCodProduto() {
 		return codProduto;
 	}
 
@@ -32,7 +42,7 @@ public class ItemPedido {
 		this.preco = preco;
 	}
 
-	public void setCodProduto(int codProduto) {
+	public void setCodProduto(Long codProduto) {
 		this.codProduto = codProduto;
 	}
